requirements.txt
httpx>=0.27.0
feedparser>=6.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
openai>=1.0.0
vaderSentiment>=3.3.2
transformers>=4.35.0
torch>=2.0.0
apscheduler>=3.10.0
